changed manager file nemes
