+++ CMakeLists.txt
@@ -13,11 +13,11 @@ add_executable(fischl
   lib/fischl.cpp
   lib/main.cpp
   lib/rawdisk.cpp
-  lib/fs/datablock_allocator.cpp
+  lib/fs/datablock_manager.cpp
   lib/fs/fs_data_types.cpp
   lib/fs/fs_resize.cpp
   lib/fs/fs.cpp
-  lib/fs/inode_allocator.cpp
+  lib/fs/inode_manager.cpp
 
 )
